Fixes #556, Renamed wrap component to kamelet-reify
