+++ camel-k-kamelet-reify/pom.xml
@@ -26,11 +26,11 @@
     <modelVersion>4.0.0</modelVersion>
     <packaging>pom</packaging>
 
-    <artifactId>camel-k-components</artifactId>
+    <artifactId>camel-k-kamelet-reify-parent</artifactId>
 
     <modules>
-        <module>camel-knative</module>
-        <module>camel-kamelet</module>
-        <module>camel-kamelet-reify</module>
+        <module>runtime</module>
+        <module>deployment</module>
     </modules>
+
 </project>
