+++ itests/camel-k-itests-kamelet-reify/src/main/resources/application.properties
