+++ itests/camel-k-itests-kamelet-reify/src/test/resources/routes.properties
