+++ support/camel-k-maven-plugin/src/main/java/org/apache/camel/k/tooling/maven/processors/CatalogProcessor3x.java
@@ -167,7 +167,7 @@ public class CatalogProcessor3x implements CatalogProcessor {
         specBuilder.putArtifact(
             new CamelArtifact.Builder()
                 .groupId("org.apache.camel.k")
-                .artifactId("camel-k-wrap")
+                .artifactId("camel-k-kamelet-reify")
                 .addScheme(new CamelScheme.Builder()
                     .id("wrap")
                     .build())
